requirements.txt
faiss-cpu
boto3
numpy